Implement improvements

+++ src/gameboy/audio.rs
@@ -1,7 +1,7 @@
 pub struct NullAudioPlayer;
 
 impl gb_core::hardware::sound::AudioPlayer for NullAudioPlayer {
-    fn play(&mut self, _output_buffer: &[f32]) {
+    fn play(&mut self, _output_buffer: &[i16]) {
         // Do nothing
     }
